Add new --create-model option
We currently have the ability to name a bundle with --name and this
is typically used in conjunction with a model of the same name. To
save time we now add the ability to autocreate a model of the same
name as --name and switch to it so that it is all ready for when we
want to deploy (and even easier if we use --run to autodeploy). If
the model already exists we switch to it and if it is already the
current model we skip and continue.