Add ability to choose K8s and charm channels on generate_bundle
This commit improves the ease of use of generate_bundle
for Kubernetes. Before it was necessary to generate a
bundle and edit the channel option for kubernetes-master
and kubernetes-worker manually. Now, you can specify
'--k8s-channel {k8s-channel}' to generate the bundle with
the option set.

Also, for the purpose of deploying Kubernetes with the
latest charm versions, if '--use-stable-charms' is not
specified, the new default will append the parameter
'--channel=edge' to the 'juju deploy' command, causing
all charms in the bundle to be deployed with their
edge channel revisions.

How to use:
- Default k8s-channel value is: 'latest/stable'.
- The latest edge charms will be used by default for
Kubernetes deployments. If you want to revert back to
using stable charms, use '--use-stable-charms' when
generating the bundle.
- There is no validation on whether the k8s-channel
you specify is valid.