Support adding resources path to bundles/overlays
When adding a path to file under resources: the path must
be relative to the path from which the deploy command is
executed i.e. that where the generate-bundle.sh is run.
To support this we allow bundles to add __RESOURCES_PATH__
which will be replaced with he patch to that generated
bundle/overlay.