Add support for octavia-diskimage-retrofit
The 19.07 charm release introduced a charm to generate
Octavia images so we add it for use by-default with an
optional flag to disable it.

When running retrofit-image action on the charm
it will attempt to find the most recent image or
alternativey one can be provided to the action using
source-image options.

Glance simplestreams is also provided as an option to
sync images to glance but is not mandatory as long as
there is an image in glance that has the properties
required to identify a valid one for conversion see
LP 1842430 for more info.