Restructure all modules to use a pipeline
Instead of having a giant generate-bundle.sh we now have it call
a pipeline of parts each with their own specific purpose i.e.

pipeline/01setup: this is for definining variables and
                  high-level init.
pipeline/02configure: this is the bulk of decision-making
                      such as overlay-selection.
pipeline/03build: calls the main generator to put it all
                  together.

Since everything now executes in the same environmemt there is no
need to support internal opts anymore so they have been removed.