Add --defaults config option
If provided this will instruct the tool to use
whatever defaults it has and not ask the caller
for REQUIRED: input.

Also tidy up the help message and check for
exclusive passthrough opts mucher earlier
on so that they immediately return.

Also fixes --list so that it always lists
contents of root state dir.