Rework opt input processing
This is a fairly large refactor that basically ensures that
all opt processing is done at the highest level. The main
reason for this is to be able to properly support exclusive
opts and overrides. It also helps to simplify the code by
storing opts that are not specific to a module in a
global dictionary that can be updated at any point.

Also moves generate-bundle.sh into common and creates a
symlink from each module since the code is the same
everywhere.