Add image format to Octavia image upload script
This change adds an image format option to the
`upload_octavia_amphora_image.sh` script to specify the image format
to use. It also adds command line option to set the OpenStack release
and the image format (which defaults to qcow2). This is useful to
force the amphora images to be in `raw` format, i.e.

        ./tools/upload_octavia_amphora_image --release stein --image-format raw

Will upload the Octavia amphora from the Stein release in raw format.

Closes-Bug: 1873307
Signed-off-by: Nicolas Bock <nicolas.bock@canonical.com>