Add options to creating an Octavia loadbalancer
The current script is not very configurable in terms of the loadbalancer
details. This change introduces several command line options which allow
the user to customize the loadbalancer created by the script.

Also enable creation of multiple loadbalancers. This was not possible
previously as the names for pool, listener, and healthmonitor would
conflict.

Signed-off-by: Nicolas Bock <nicolas.bock@canonical.com>