Don't clobber data-port when re-running
Bundles that depend on --dvr require a data-port
which results in a user input-request. If
data-port has been set and you re-run generate-bundle
it will ask for a data-port value and default to
empty. This patch allows for any existing charm value
to be re-applied so as not to clobber

Closes-Bug: 1871565