Remove requirement to confirm machine poweron
While it makes sense to ask confirmation when shutting down
the VMs for a model, powering on does not. This is particularly
annoying when STS is a bit slow and after starting the model
and switch to other tasks, you came back later just to find
that your model still not up.