Fix the the exit code of generate-bundle.sh
The final exit code of this script is non-zero when it's success.

When grep fails (i.e. no unbound variables), it exits immediately
because of -e option. || ensures that grep failure doesn't exit the
script and also explicitly exit with 0. Otherwise, "last command"'s
(= grep) return code will be used for exit code which is non-zero.

Signed-off-by: Ponnuvel Palaniyappan <pponnuvel@gmail.com>