Don't exclude namespaces charms from revision-info
Most charms release to cs:<charm> as their stable whereas
others use cs:~<team>/<charm> so we need to allow that
when using --release-info.

One remaining caveat here is that if the revision-info uses
a namespaced charm but the template does not then it will
not be matched. That will either require a manual edit to
fix (--replay) or we need to update the template in question
to use the namespaced charm.

Closes-Bug: 1881158