Add parsing of `juju status` output for revision info
This change adds logic to the parse helper script to be able to parse
the output of `juju status` to extract charm revision information. The
change reuses the `--revision-info` command line argument. The file
given is first parsed assuming it contains an exported bundle in YAML
format. If that fails the file is reparsed assuming it contains the
output of `juju status`.

Signed-off-by: Nicolas Bock <nicolas.bock@canonical.com>