Update __OCTAVIA_RETROFIT_UCA__ for non UCA deployments
__OCTAVIA_RETROFIT_UCA__ is updated only for UCA release deployments.
And the function get_uca_release returns nothing if release is not
mentioned in generate_bundle.

New function get_ost_release introduced that returns the release name
specified as option for generate_bundle. If release option is not
specified, the release for corresponding series is returned. This
ensures param __OCTAVIA_RETROFIT_UCA__ is updated properly.

Closes-Bug: 1888100