Use the loadbalancer name as base for its children
In order to deploy multiple load balancers using the
`create_octavia_lb.sh` script it is necessary to prefix or make unique
in some way the names of the pools, listeners, and health monitors.
This change uses the value passed to the script using the `--name`
command line option as prefix for the names of any subsequent load
balancer components.

This change also adjusts some `openstack` CLI command line flags to
use the long version for readability.

Signed-off-by: Nicolas Bock <nicolas.bock@canonical.com>