Speed-up application of version-info
If the yaml provided to --version info is large, it
can result in a very large number of unnecessary sed
executions that can really slow down the time it
takes to complete generating a bundle. This patch
reduces this time significantly by skipping the sed
if the it isnt needed.