Use default-tenant-network-type to specify default
A previous commit changed the default overlay-network-type from 'gre' to
'vxlan'. Unfortunately this setting restricts what kinds of networks work in
the cloud, and changing this on a running cloud (which happens from re-running
the same generate-bundle to add something extra) will break the existing
networks and this is not simple to diagnose.

Hence we always set 'overlay-network-type' to 'gre vxlan geneve' to always
allow both and switch to using default-tenant-network-type to specify the
default.

Also update some of the scripts that relied on overlay-network-type having a
single value.