Add command line option to `novarc`
This change adds the `--service` command line option to `novarc`. This
will set the OpenStack credentials in the shell environment to a
specific service as opposed to the admin user. For example

    $ . novarc --service octavia

will set the credentials such that one can now issue OpenStack CLI
commands as the octavia user and do things such as

    $ openstack keypair list
    +------------------+-------------------------------------------------+
    | Name             | Fingerprint                                     |
    +------------------+-------------------------------------------------+
    | amphora-backdoor | 50:9c:e3:9e:06:b5:b6:40:9c:a0:65:30:73:fa:3e:19 |
    +------------------+-------------------------------------------------+

Signed-off-by: Nicolas Bock <nicolas.bock@canonical.com>