Fix config-based ssl novarc
When using config based ssl (i.e. not vault) the certs
are stored under a path named with the model deployed
and as such the old path of openstack/ssl no longer
works so we remove refs to that and use a selection
method to choose the right one if more than one
exists.