Only try to power-on SHUTOFF VMs
Sometimes not all machines are powered-off in a model and running the
script returns an error due to the fact that it is trying to power-on
a VM that is not in a 'SHUTOFF' state.
This fix aims to improve the script so that only machines in SHUTOFF
state in the model are powered on.

Signed-off-by: David Negreira <david.negreira@canonical.com>