Add SSL configuration for placement API overlay
This patch aims to add the option to configure SSL on the placement API unit.

Since there isa no HA option for placement API, and since the other units
are using their HA overlay with 1 unit, I have added that as well for the
placement API.

I have taken they keystone-ha and SSL configuration as reference.

Signed-off-by: David Negreira <david.negreira@canonical.com>