Add script to look at logs from multiple units
The `juju-lnav` script is a helper script to call `lnav` (needs to be
installed via package or snap) on multiple machines, applications, or
units. For example,

    $ ./tools/juju-lnav octavia:/var/log/octavia/*.log{,1}

Will open the current and the first rotated log files from all of the
`octavia` units.

Likewise

    $ ./tools/juju-lnav octavia:/var/log/octavia/*.log \
        nova-cloud-controller:/var/log/nova/*.log

Will load the logs on all `octavia` and all `nova-cloud-controller`
units.

For more information on how to use `lnav` please visit lnav.org.

Signed-off-by: Nicolas Bock <nicolas.bock@canonical.com>