Add cinder-volume and cinder-purestorage overlays
Add support for deploying cinder-purestorage

Add support for deploying a separate cinder-volume service. Though not
strictly required in stsstack, this is to replicate customer
environments where the LXD cinder service cannot connect to iSCSI
volumes to clone images to them (as iSCSI is not namespaced), so a
separate cinder-volume service is deployed in KVM or on metal. Commonly
used with purestorage or other supported iSCSI SANs.

Also ensures we specify cinder:identity-service which is required as of
a recent change to the cinder charm.