Add lunar/antelope support
Due to some checks being done by alphabetical order, it would break
how we would assign and ensure that minimal versions of certain charms
are installed in certain openstack versions.

We fix this by putting all the relevant openstack versions in an
ordered array so that we can compare numerically the openstack
versions between eachother.

Closes: #112

Co-authored-by: David Negreira <david.negreira@canonical.com>
Co-authored-by: Edward Hope-Morley <edward.hope-morley@canonical.com>

Signed-off-by: David Negreira <david.negreira@canonical.com>