Move vault unseal before ./configure
The tutorial doc, which prints a message after
generate-bundle.sh is run, has ./configure as the
first step. But if you do that there is a complaint
that ./tools/vault-unseal-and-authorise.sh has not
been run yet. A similar message is printed if you
source novarc.

Move it to before ./configure since it must be
run first.