Set image properties less aggressively
The current implementation parallelizes the setting of image properties across the images and the properties. This can lead to some of the properties getting "lost" and not getting set. Missing image properties will break the octavia-diskimage-retrofit charm when one of those images is used as the base for the amphora image.

The `juju run` action has a `--wait` parameter which defaults to `0s` which really means 1 minute. On PS6 this timeout is too aggressive and the configure and retrofit actions time out.

Signed-off-by: Nicolas Bock <nicolas.bock@canonical.com>