Noble-Caracal: Use OVN charms from latest/edge
Update the Noble channel mappings for OVN to use latest/edge.

Jammy-Caracal is supported by OVN charms in the 24.03 channel, but
Noble-Caracal fails to deploy because the binary builds on that channel
only support Jammy and not Noble.

Supporting Noble-Caracal in the charms officially may not be coming at
all, but support specifically for the purpose of testing package SRUs
has been released on the latest/edge channel.