Change juju-lnav to bail on failed ssh
* Remove install_ssh_key
* Add check_ssh_key and exit if unable to connect

Previously juju-lnav attempted to copy the keys
'~/.ssh/id_rsa.pub' and '~/testkey.pub' to a server
when connecting to a machine, if an initial ssh test failed.
This could lead to confusing errors if either of the above keys
was missing, and wouldn't necessarily fix connectivity issues.
Also, lnav doesn't make it especially clear what's going on
if it is unable to ssh.

This commit changes the script to bail out if it can't establish
and ssh connection with the target machine, and instead
prints instructions to the user letting them know they
should add an ssh key to the server and try again.