juju-lnav: trim .priv when guessing ssh pub key
The logic added in #322 to guess the public key path from the private
key path turned out to be a little to naive, i.e. it just added .pub to
the end. Other scripts create the pair testkey.priv and testkey.pub so a
smarter guess would probably be to strip .priv from the end if it's
there first and then add .pub.

Fixes #328

Signed-off-by: Zachary Raines <zachary.raines@canonical.com>

Rename guess_ssh_key to guess_pub_key

Fix wrong name