Refactor novarc into tool and shim
The old openstack/novarc script has been moved
mostly unchanged into openstack/tools/construct_novarc.sh
with the difference that instead of directly setting the
env vars, it outputs a valid novarc file that can be sourced.

The openstack/novarc script has then been refactored to
simply source the output of the tool. This guarantees
compatibility with bash and zsh. As an added benefit,
the novarc can be saved while working on model,
so that it can be sourced quickly without rerunning the
juju commands to discover things.

Signed-off-by: Zachary Raines <zachary.raines@canonical.com>