Refactor novarc into tool and shim
The old openstack/novarc script has been moved
mostly unchanged into openstack/tools/construct_novarc.sh
with the difference that instead of directly setting the
env vars, it outputs a valid novarc file that can be sourced,
and a few code quality improvements noted below.

The openstack/novarc script has then been refactored to
simply source the output of the tool. This guarantees
compatibility with bash and zsh. As an added benefit,
the novarc can be saved while working on model,
so that it can be sourced quickly without rerunning the
juju commands to discover things.

Changes to novarc code:

* Don't try to trap KILL in novarc script
    This won't work anyway.

* Dont -rf on cleanup of juju_json_status_cache
    It's not needed makes it less likely that something SCARY happens.

* Improve quoting in novarc script for safety

* Fix incorrect if statement in novarc

    Previously this checked the output of grep -q, which is always empty,
    instead of the exit code.

* Make array indexing explicit in novarc

* Fix erroneous return in novarc script

    We can't return from the top level. Should be an exit instead.

* Convert tests to [[ ]] in construct_novarc.sh

Signed-off-by: Zachary Raines <zachary.raines@canonical.com>